Order dev

+++ EducationApp/EducationApp.BusinessLogicLayer/Models/ResponseModels/Order/OrderItemModel.cs
@@ -6,10 +6,10 @@ namespace EducationApp.BusinessLogicLayer.Models.ResponseModels.Order
 {
     public class OrderItemModel
     {
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public Currency Currency { get; set; }
-        public decimal Count { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int Count { get; set; }
+        public long UnitPrice { get; set; }
         public Guid PrintingEditionId { get; set; }
         public Guid OrderId { get; set; }
     }
