+++ EducationApp/EducationApp.DataAccessLayer/Entities/OrderItem.cs
@@ -6,10 +6,10 @@ namespace EducationApp.DataAccessLayer.Entities
 {
     public class OrderItem : BaseEntity
     {
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public Currency Currency { get; set; }
-        public decimal Count { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int Count { get; set; }
+        public long UnitPrice { get; set; }
         public Guid PrintingEditionId { get; set; }
         public PrintingEdition PrintingEdition { get; set; }
         public Guid OrderId { get; set; }
