+++ EducationApp/EducationApp.DataAccessLayer/Entities/PrintingEdition.cs
@@ -8,7 +8,7 @@ namespace EducationApp.DataAccessLayer.Entities
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public long Price { get; set; }
         public string Type { get; set; }
         public Status Status { get; set; }
         public Currency Currency { get; set; }
